fix(ci): 补齐 clippy 存量告警，修 macOS 与 Windows 双侧门禁
CI 自 8-13 起连续红 8 次，根因是两件事叠加，不是代码劣化：

1. b2cdaffc 把 macOS 门禁从 `cargo clippy --workspace` 收紧为
   `--all-targets -- -D warnings`。那条提交自己就预告过「此前从未按
   --all-targets 跑过，首次触发可能暴露 macOS 侧的存量警告」——本次
   修的主体正是这批存量。
2. Rust 1.97 新增的 manual_is_multiple_of、unnecessary_to_owned 等
   lint，让既有代码凭空违规。

★ clippy 首个 crate 失败即停止调度，CI 每次只报得出一个错误，实际共
22 个、分五层。只跟着 CI 日志修就会陷入「推一次、等两分钟、再报一个」
的循环。本地 --keep-going 逐层剥完再一次修完。

按性质分三类：

- cfg 不对称（多数）。无条件的 use 里有一部分符号只被 cfg(windows) 的
  函数消费，darwin 下即成死代码。修法是给 import 与符号加 cfg 门，
  ⛔不能删——clippy --fix 只看当前 target，会把 Windows 侧需要的直接
  删掉，换来另一个平台编译失败。wind-bridge 的 security 模块（SDDL 纯
  属 Windows 概念、消费者全在 cfg(windows) 内）整个下沉到 cfg(windows)。
- mock 刻意保留的形状。wind-ui/sys.rs 的非 Windows mock 是照抄 windows
  crate 的签名（PascalCase 函数名、VIRTUAL_KEY、unsafe fn 无 # Safety、
  Result<_, ()>），改成「更 Rust」的写法就失去了本模块存在的意义——
  调用点在两个平台上必须一字不差。故用模块级 allow 并写明理由。
- 真实缺陷。manager_macos / global_hotkey_macos 的 4 处多余 cast 与一处
  复杂类型（抽 HotkeyActions alias），照实修。

验证：fmt --check、clippy(x86_64-pc-windows-msvc)、
clippy(aarch64-apple-darwin)、test --workspace（词库测试族耗时 > 0，
非静默跳过）、check -p wind-mobile(aarch64-linux-android) 与两条依赖树
断言，均通过。

⚠ wind-webdata / wind-transfer 含 zstd-sys、bzip2-sys、lzma-sys，本机
无 macOS C 交叉工具链，这两个 crate 的 darwin lint 未能本地覆盖，留给
CI 把关。

+++ wind_input/crates/wind-ui/src/manager_macos.rs
@@ -522,7 +522,7 @@ impl Forwarder {
             return None;
         }
         let f = self.win.render_frame()?;
-        Some((f.buf, f.width as u32, f.height as u32))
+        Some((f.buf, f.width, f.height))
     }
 
     /// 请 `.app` 截某个原生浮窗存盘。文件名在此定（与 Windows 侧同一格式），
